Unified Sidekiq patch file
